refactor(repo): modularize application layer into subpackages (serialization)
