+++ src/aigora/curriculum_graph/application/serialization/serializer_errors.py
